MAS-246 extract change calculation
